Main페이지 삭제 및 MainActivity 변경
